+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 
 ### Improvements
 
+- instance: show migrated to egoscale v3
 - SKS: more related commands are migrated to egoscale v3
 - dbaas: move all commands to egoscale v3
 - EIP: move to egoscale v3 #684